+++ core/src/main/java/org/hisp/dhis/android/core/relationship/RelationshipHandlerImpl.java
@@ -116,7 +116,7 @@ final class RelationshipHandlerImpl implements RelationshipHandler {
                 RelationshipStore.create(databaseAdapter),
                 RelationshipItemStoreImpl.create(databaseAdapter),
                 RelationshipItemHandler.create(databaseAdapter),
-                RelationshipItemElementStoreSelector.create(databaseAdapter),
+                RelationshipItemElementStoreSelectorImpl.create(databaseAdapter),
                 new RelationshipDHISVersionManager(versionManager)
         );
     }
