+++ core/src/main/java/org/hisp/dhis/android/core/relationship/RelationshipItemElementStoreSelectorImpl.java
@@ -1,0 +1,73 @@
+/*
+ * Copyright (c) 2017, University of Oslo
+ *
+ * All rights reserved.
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ * Neither the name of the HISP project nor the names of its contributors may
+ * be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package org.hisp.dhis.android.core.relationship;
+
+import org.hisp.dhis.android.core.common.StoreWithState;
+import org.hisp.dhis.android.core.data.database.DatabaseAdapter;
+import org.hisp.dhis.android.core.enrollment.EnrollmentStore;
+import org.hisp.dhis.android.core.enrollment.EnrollmentStoreImpl;
+import org.hisp.dhis.android.core.event.EventStore;
+import org.hisp.dhis.android.core.event.EventStoreImpl;
+import org.hisp.dhis.android.core.trackedentity.TrackedEntityInstanceStore;
+import org.hisp.dhis.android.core.trackedentity.TrackedEntityInstanceStoreImpl;
+
+final class RelationshipItemElementStoreSelectorImpl implements RelationshipItemElementStoreSelector {
+
+    private final TrackedEntityInstanceStore trackedEntityInstanceStore;
+    private final EnrollmentStore enrollmentStore;
+    private final EventStore eventStore;
+
+    private RelationshipItemElementStoreSelectorImpl(
+            TrackedEntityInstanceStore trackedEntityInstanceStore,
+            EnrollmentStore enrollmentStore,
+            EventStore eventStore) {
+
+        this.trackedEntityInstanceStore = trackedEntityInstanceStore;
+        this.enrollmentStore = enrollmentStore;
+        this.eventStore = eventStore;
+    }
+
+    @Override
+    public StoreWithState getElementStore(RelationshipItem item) {
+        if (item.hasTrackedEntityInstance()) {
+            return trackedEntityInstanceStore;
+        } else if (item.hasEnrollment()) {
+            return enrollmentStore;
+        } else {
+            return eventStore;
+        }
+    }
+
+    public static RelationshipItemElementStoreSelectorImpl create(DatabaseAdapter databaseAdapter) {
+        return new RelationshipItemElementStoreSelectorImpl(
+                new TrackedEntityInstanceStoreImpl(databaseAdapter),
+                new EnrollmentStoreImpl(databaseAdapter),
+                new EventStoreImpl(databaseAdapter)
+        );
+    }
+}
