+++ core/src/main/java/org/hisp/dhis/android/core/relationship/RelationshipRepositoryImpl.java
@@ -178,7 +178,7 @@ final class RelationshipRepositoryImpl extends ReadOnlyCollectionRepositoryImpl<
                 Relationship.factory,
                 relationshipHandler,
                 RelationshipItemStoreImpl.create(databaseAdapter),
-                RelationshipItemElementStoreSelector.create(databaseAdapter),
+                RelationshipItemElementStoreSelectorImpl.create(databaseAdapter),
                 new RelationshipItemPojoBuilder()
         );
     }
